fix: pop_bkac(), back(), front()

+++ include/ft.hpp
@@ -13,7 +13,7 @@ namespace ft
 	std::vector<std::string> split(std::string const & str);
 	std::vector<std::string> split(std::string const & str, std::string const & delim);
 	int toInt(std::string const & str, int unit);
-	std::string toString(ssize_t nbr, int unit);
+	std::string toString(long long nbr, int unit);
 }
 
 #endif
